7-Zip Version 2.30 Beta 24
  SDK/Windows/Synchronization.h
  SDK/Windows/Synchronization.cpp
    - some changes.

+++ Project/Archiver/Format/BZip2/Handler.h
@@ -30,7 +30,7 @@ END_COM_MAP()
 DECLARE_NOT_AGGREGATABLE(CHandler)
 
 DECLARE_REGISTRY(CHandler, TEXT("SevenZip.FormatBZip2.1"), 
-    TEXT("SevenZip.FormatBZip2"), 0, THREADFLAGS_APARTMENT)
+    TEXT("SevenZip.FormatBZip2"), UINT(0), THREADFLAGS_APARTMENT)
 
   STDMETHOD(Open)(IInStream *aStream, 
       const UINT64 *aMaxCheckStartPosition,
