+++ Project/Archiver/Format/Zip/Handler.h
@@ -36,7 +36,7 @@ END_COM_MAP()
 DECLARE_NOT_AGGREGATABLE(CZipHandler)
 
 DECLARE_REGISTRY(CZipHandler, TEXT("SevenZip.FormatZip.1"), 
-  TEXT("SevenZip.FormatZip"), 0, THREADFLAGS_APARTMENT)
+  TEXT("SevenZip.FormatZip"), UINT(0), THREADFLAGS_APARTMENT)
 
   STDMETHOD(Open)(IInStream *aStream, 
       const UINT64 *aMaxCheckStartPosition,
