+++ Project/Compress/Convert/Branch/Coder.h
@@ -34,7 +34,7 @@ BEGIN_COM_MAP(C ## Name) \
 END_COM_MAP() \
 DECLARE_NOT_AGGREGATABLE(C ## Name) \
   DECLARE_REGISTRY(C ## Name, TEXT("Compress.ConvertBranch.1"), \
-  TEXT("Compress.ConvertBranch"), 0, THREADFLAGS_APARTMENT) \
+  TEXT("Compress.ConvertBranch"), UINT(0), THREADFLAGS_APARTMENT) \
   STDMETHOD(Code)(ISequentialInStream *inStream, \
       ISequentialOutStream *outStream, const UINT64 *inSize, const UINT64 *outSize, \
       ICompressProgressInfo *progress); \
