+++ Project/Compress/LZ/Deflate/Decoder.h
@@ -86,7 +86,7 @@ public:
   // DECLARE_NO_REGISTRY()
 
   DECLARE_REGISTRY(CEncoder, TEXT("Compress.DeflateDecoder.1"), 
-  TEXT("Compress.DeflateDecoder"), 0, THREADFLAGS_APARTMENT)
+  TEXT("Compress.DeflateDecoder"), UINT(0), THREADFLAGS_APARTMENT)
 
   STDMETHOD(CodeReal)(ISequentialInStream *anInStream,
       ISequentialOutStream *anOutStream, const UINT64 *anInSize, const UINT64 *anOutSize,
