+++ Project/Compress/LZ/Deflate/Encoder.h
@@ -151,7 +151,7 @@ public:
 
   // DECLARE_NO_REGISTRY()
   DECLARE_REGISTRY(CEncoder, TEXT("Compress.DeflateEncoder.1"), 
-  TEXT("Compress.DeflateEncoder"), 0, THREADFLAGS_APARTMENT)
+  TEXT("Compress.DeflateEncoder"), UINT(0), THREADFLAGS_APARTMENT)
 
   HRESULT CodeReal(ISequentialInStream *anInStream,
       ISequentialOutStream *anOutStream, const UINT64 *anInSize, const UINT64 *anOutSize,
