+++ Project/Compress/LZ/MatchFinder/HashChain/HCMF.h
@@ -81,7 +81,8 @@ END_COM_MAP()
 
 DECLARE_NOT_AGGREGATABLE(CMatchFinderHC)
 
-DECLARE_REGISTRY(CMatchFinderHC, kIDStringFull TEXT(".1"), kIDStringFull, 0, THREADFLAGS_APARTMENT)
+DECLARE_REGISTRY(CMatchFinderHC, kIDStringFull TEXT(".1"), kIDStringFull, 
+	UINT(0), THREADFLAGS_APARTMENT)
 
   STDMETHOD(Init)(ISequentialInStream *aStream);
   STDMETHOD_(void, ReleaseStream)();
